test: split turn memory white-box suites
