+++ crates/agent-engine/src/runtime/turn_executor.rs
@@ -1873,816 +1873,12 @@ description: {description}
 
     mod turn_execution;
 
-    #[tokio::test]
-    async fn test_run_turn_refreshes_memory_surfaces_when_tool_batch_ends_turn() {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-
-        let mut state = create_test_state_with_provider(ToolCallMockProvider::new(
-            vec![ToolCall {
-                id: Some("call_1".to_string()),
-                name: "request_confirmation".to_string(),
-                arguments: json!({}),
-            }],
-            "",
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir.clone());
-        state
-            .turn_state
-            .set_turn_activity(TurnActivityState::Running);
-
-        let cancel = CancellationToken::new();
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text("Test input")]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert!(matches!(result.unwrap(), TurnExecutionOutcome::Finished));
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::Error { message, .. } if message == "Invalid confirmation request."
-        )));
-        assert!(memory_dir.join("handoffs").join("LATEST.md").exists());
-        assert!(memory_dir.join("episodic").exists());
-        assert!(
-            std::fs::read_dir(memory_dir.join("daily"))
-                .unwrap()
-                .next()
-                .is_some()
-        );
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_promotes_direct_user_fact_when_tool_batch_ends_turn() {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-
-        let mut state = create_test_state_with_provider(ToolCallMockProvider::new(
-            vec![ToolCall {
-                id: Some("call_1".to_string()),
-                name: "request_confirmation".to_string(),
-                arguments: json!({}),
-            }],
-            "",
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir.clone());
-        state.core_config.memory.enabled = true;
-        state
-            .turn_state
-            .set_turn_activity(TurnActivityState::Running);
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text("My name is Morris.")]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert!(matches!(result.unwrap(), TurnExecutionOutcome::Finished));
-        let user_memory_before =
-            std::fs::read_to_string(memory_dir.join("USER.md")).unwrap_or_else(|_| String::new());
-        assert!(!user_memory_before.contains("Name: Morris"));
-        assert_eq!(run_deferred_runtime_actions(&mut state).await, 1);
-
-        let user_memory = std::fs::read_to_string(memory_dir.join("USER.md")).unwrap();
-        assert!(user_memory.contains("Name: Morris"));
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_defers_memory_promotion_until_after_completion() {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-
-        let mut state = create_test_state_with_provider(FailOnMemoryPromotionProvider {
-            content: "Done.".to_string(),
-        });
-        state.core_config.memory.store_dir = Some(memory_dir);
-        state.core_config.memory.enabled = true;
-
-        let cancel = CancellationToken::new();
-        let mut events = Vec::new();
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text("My name is Morris.")]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert!(matches!(result.unwrap(), TurnExecutionOutcome::Finished));
-        assert!(
-            events
-                .iter()
-                .any(|event| matches!(event, Event::TurnCompleted { .. }))
-        );
-        assert_eq!(state.turn_state.drain_deferred_runtime_actions().len(), 1);
-    }
-
-    struct SlowTool {
-        delay: tokio::time::Duration,
-    }
-
-    impl Tool for SlowTool {
-        fn name(&self) -> &str {
-            "slow_tool"
-        }
-
-        fn description(&self) -> &str {
-            "Slow tool used to test cancellation."
-        }
-
-        fn parameters_schema(&self) -> serde_json::Value {
-            json!({
-                "type": "object",
-                "properties": {}
-            })
-        }
-
-        fn execute(&self, _arguments: serde_json::Value, _ctx: &ToolContext) -> ToolResult {
-            let delay = self.delay;
-            Box::pin(async move {
-                tokio::time::sleep(delay).await;
-                Ok(json!({ "ok": true }))
-            })
-        }
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_cancelled_tool_batch_does_not_refresh_memory_surfaces() {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-
-        let mut tools = ToolRegistry::new();
-        tools.register(SlowTool {
-            delay: tokio::time::Duration::from_millis(50),
-        });
-        let mut state = create_test_state_with_provider_and_tools(
-            ToolCallMockProvider::new(
-                vec![ToolCall {
-                    id: Some("call_1".to_string()),
-                    name: "slow_tool".to_string(),
-                    arguments: json!({}),
-                }],
-                "",
-            ),
-            tools,
-        );
-        state.core_config.memory.store_dir = Some(memory_dir.clone());
-        state
-            .turn_state
-            .set_turn_activity(TurnActivityState::Running);
-
-        let cancel = CancellationToken::new();
-        let cancel_for_task = cancel.clone();
-        tokio::spawn(async move {
-            tokio::time::sleep(tokio::time::Duration::from_millis(10)).await;
-            cancel_for_task.cancel();
-        });
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text("Test input")]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert!(matches!(result.unwrap(), TurnExecutionOutcome::Finished));
-        assert!(events.iter().any(|event| matches!(
-            event,
-            Event::TurnCompleted { summary: Some(summary) }
-                if summary == "Task cancelled by user"
-        )));
-        assert!(!memory_dir.join("handoffs").join("LATEST.md").exists());
-        assert!(!memory_dir.join("episodic").exists());
-        assert!(!memory_dir.join("daily").exists());
-    }
-
-    #[tokio::test]
-    #[allow(
-        clippy::field_reassign_with_default,
-        reason = "the test highlights only the memory-refresh fields that define this scenario"
-    )]
-    async fn test_run_turn_tool_loop_guard_refreshes_memory_surfaces_before_completion_event() {
-        let temp = TempDir::new().unwrap();
-        let memory_dir = temp.path().join("memory-store");
-        let generate_calls = Arc::new(AtomicUsize::new(0));
-        let provider = SequenceMockProvider::new(
-            vec![
-                GenerationResponse {
-                    content: String::new(),
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: Vec::new(),
-                    tool_calls: vec![ToolCall {
-                        id: Some("call-1".to_string()),
-                        name: "update_plan".to_string(),
-                        arguments: json!({
-                            "explanation": "Loop 1",
-                            "items": [{"id": "1", "content": "Step 1", "status": "in_progress"}]
-                        }),
-                    }],
-                    usage: None,
-                    finish_reason: None,
-                    warnings: Vec::new(),
-                    provider_response_id: None,
-                    provider_response_status: None,
-                },
-                GenerationResponse {
-                    content: String::new(),
-                    thinking: None,
-                    thinking_signature: None,
-                    redacted_thinking: Vec::new(),
-                    tool_calls: vec![ToolCall {
-                        id: Some("call-2".to_string()),
-                        name: "update_plan".to_string(),
-                        arguments: json!({
-                            "explanation": "Loop 2",
-                            "items": [{"id": "2", "content": "Step 2", "status": "in_progress"}]
-                        }),
-                    }],
-                    usage: None,
-                    finish_reason: None,
-                    warnings: Vec::new(),
-                    provider_response_id: None,
-                    provider_response_status: None,
-                },
-            ],
-            Arc::clone(&generate_calls),
-        );
-        let mut state = create_test_state_with_provider(provider);
-        state.core_config.memory.store_dir = Some(memory_dir.clone());
-        state.runtime_config.max_tool_loops = 2;
-
-        let cancel = CancellationToken::new();
-        let mut saw_handoff_before_completion = false;
-        let mut emit = |event: Event| {
-            if matches!(
-                event,
-                Event::TurnCompleted {
-                    summary: Some(ref summary)
-                } if summary == "Tool loop stopped by loop guard"
-            ) {
-                saw_handoff_before_completion = memory_dir.join("handoffs/LATEST.md").exists();
-            }
-            async {}
-        };
-
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text(
-                "Run until the loop guard stops you.",
-            )]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert!(matches!(result.unwrap(), TurnExecutionOutcome::Finished));
-        assert_eq!(generate_calls.load(Ordering::SeqCst), 2);
-        assert_eq!(state.turn_state.drain_deferred_runtime_actions().len(), 1);
-        assert!(saw_handoff_before_completion);
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_confirmation_includes_active_skill_permission_hints() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let skill_dir = definition_root.join("skills/release-check");
-        std::fs::create_dir_all(&skill_dir).unwrap();
-        std::fs::write(
-            skill_dir.join("SKILL.md"),
-            r#"---
-name: Release Check
-description: Review risky release actions
----
-
-# Instructions
-Use this skill when asked.
-"#,
-        )
-        .unwrap();
-        std::fs::write(
-            skill_dir.join("skill.yaml"),
-            r#"
-runtime:
-  permission_hints:
-    - "May require write approval."
-"#,
-        )
-        .unwrap();
-
-        let mut state = create_test_state_with_provider(ToolCallMockProvider::new(
-            vec![ToolCall {
-                id: Some("call_1".to_string()),
-                name: "request_confirmation".to_string(),
-                arguments: json!({
-                    "checkpoint_type": "test",
-                    "summary": "Confirm risky action"
-                }),
-            }],
-            "",
-        ));
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-        let cancel = CancellationToken::new();
-
-        let mut events = vec![];
-        let mut emit = |event: Event| {
-            events.push(event);
-            async {}
-        };
-
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text(
-                "please use $release-check for this task",
-            )]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-
-        let confirmation = events.into_iter().find_map(|event| match event {
-            Event::Yield {
-                kind: alan_agent_protocol::YieldKind::Confirmation,
-                payload,
-                ..
-            } => Some(payload),
-            _ => None,
-        });
-        let confirmation = confirmation.expect("expected confirmation yield");
-        let hints = confirmation["details"]["skill_permission_hints"]
-            .as_array()
-            .cloned()
-            .unwrap();
-
-        assert_eq!(hints.len(), 1);
-        assert_eq!(hints[0]["skill_id"], "release-check");
-        assert_eq!(
-            hints[0]["permission_hints"][0],
-            "May require write approval."
-        );
-    }
-
-    struct RecordingToolCallProvider {
-        tool_calls: Vec<ToolCall>,
-        content: String,
-        seen_system_prompts: Arc<std::sync::Mutex<Vec<String>>>,
-    }
-
-    impl RecordingToolCallProvider {
-        fn new(
-            tool_calls: Vec<ToolCall>,
-            content: impl Into<String>,
-            seen_system_prompts: Arc<std::sync::Mutex<Vec<String>>>,
-        ) -> Self {
-            Self {
-                tool_calls,
-                content: content.into(),
-                seen_system_prompts,
-            }
-        }
-
-        fn record_system_prompt(&self, request: &GenerationRequest) {
-            if let Some(system_prompt) = request.system_prompt.as_ref() {
-                self.seen_system_prompts
-                    .lock()
-                    .unwrap()
-                    .push(system_prompt.clone());
-            }
-        }
-    }
-
-    #[async_trait]
-    impl LlmProvider for RecordingToolCallProvider {
-        async fn generate(
-            &mut self,
-            request: GenerationRequest,
-        ) -> anyhow::Result<GenerationResponse> {
-            self.record_system_prompt(&request);
-            if let Some(response) = maybe_memory_promotion_response(&request) {
-                return Ok(response);
-            }
-            Ok(GenerationResponse {
-                content: self.content.clone(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: self.tool_calls.clone(),
-                usage: None,
-                finish_reason: None,
-                warnings: Vec::new(),
-                provider_response_id: None,
-                provider_response_status: None,
-            })
-        }
-
-        async fn chat(&mut self, _system: Option<&str>, _user: &str) -> anyhow::Result<String> {
-            Ok(format!("mock: {}", self.content))
-        }
-
-        async fn generate_stream(
-            &mut self,
-            request: GenerationRequest,
-        ) -> anyhow::Result<tokio::sync::mpsc::Receiver<StreamChunk>> {
-            self.record_system_prompt(&request);
-            if let Some(response) = maybe_memory_promotion_response(&request) {
-                return Ok(response_stream(response));
-            }
-            Ok(response_stream(GenerationResponse {
-                content: self.content.clone(),
-                thinking: None,
-                thinking_signature: None,
-                redacted_thinking: Vec::new(),
-                tool_calls: self.tool_calls.clone(),
-                usage: None,
-                finish_reason: None,
-                warnings: Vec::new(),
-                provider_response_id: None,
-                provider_response_status: None,
-            }))
-        }
-
-        fn provider_name(&self) -> &'static str {
-            "recording_tool_call_mock"
-        }
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_includes_runtime_recall_bundle_for_identity_query() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let memory_dir = definition_root.join("memory-store");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        std::fs::write(
-            memory_dir.join("USER.md"),
-            "# User Memory\n- Favorite runtime marker: ALAN_IDENTITY_RECALL\n",
-        )
-        .unwrap();
-
-        let seen_system_prompts = Arc::new(std::sync::Mutex::new(Vec::new()));
-        let mut state = create_test_state_with_provider(RecordingToolCallProvider::new(
-            Vec::new(),
-            "ALAN_IDENTITY_RECALL",
-            seen_system_prompts.clone(),
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir);
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text(
-                "What is my favorite runtime marker?",
-            )]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-
-        let system_prompts = seen_system_prompts.lock().unwrap();
-        let request_prompt = system_prompts
-            .iter()
-            .find(|prompt| prompt.contains("## Runtime Recall Bundle"))
-            .expect("expected runtime recall bundle prompt");
-        assert!(request_prompt.contains("## Runtime Recall Bundle"));
-        assert!(request_prompt.contains("/memory/USER.md"));
-        assert!(request_prompt.contains("ALAN_IDENTITY_RECALL"));
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_includes_runtime_recall_bundle_for_continuity_query() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let memory_dir = definition_root.join("memory-store");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        std::fs::write(
-            memory_dir.join("handoffs/LATEST.md"),
-            "# Latest Handoff\n- Continuity marker: ALAN_CONTINUITY_RECALL\n",
-        )
-        .unwrap();
-        std::fs::create_dir_all(memory_dir.join("episodic/2026/04/15")).unwrap();
-        std::fs::write(
-            memory_dir.join("episodic/2026/04/15/process-1.md"),
-            "# Agent Process Activity\n- Continuity marker: ALAN_CONTINUITY_RECALL\n",
-        )
-        .unwrap();
-
-        let seen_system_prompts = Arc::new(std::sync::Mutex::new(Vec::new()));
-        let mut state = create_test_state_with_provider(RecordingToolCallProvider::new(
-            Vec::new(),
-            "ALAN_CONTINUITY_RECALL",
-            seen_system_prompts.clone(),
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir);
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text(
-                "What was the previous Agent Process doing?",
-            )]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-
-        let system_prompts = seen_system_prompts.lock().unwrap();
-        let request_prompt = system_prompts
-            .iter()
-            .find(|prompt| prompt.contains("## Runtime Recall Bundle"))
-            .expect("expected runtime recall bundle prompt");
-        assert!(request_prompt.contains("## Runtime Recall Bundle"));
-        assert!(request_prompt.contains("/memory/handoffs/LATEST.md"));
-        assert!(request_prompt.contains("/memory/episodic/2026/04/15/process-1.md"));
-        assert!(request_prompt.contains("ALAN_CONTINUITY_RECALL"));
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_includes_runtime_recall_bundle_for_recent_query_fallback() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let memory_dir = definition_root.join("memory-store");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        std::fs::create_dir_all(memory_dir.join("episodic/2026/04/16")).unwrap();
-        for index in 1..=4 {
-            std::fs::write(
-                memory_dir.join(format!("topics/recent-match-{index}.md")),
-                format!("# Topic Note\nwe did document topic match {index}\n"),
-            )
-            .unwrap();
-        }
-        std::fs::write(
-            memory_dir.join("daily/2026-04-16.md"),
-            "## 2026-04-16\nALAN_RECENT_RECALL\n",
-        )
-        .unwrap();
-        for index in 1..=4 {
-            std::fs::write(
-                memory_dir.join(format!("episodic/2026/04/16/process-{index}.md")),
-                format!("# Agent Process Activity\nALAN_RECENT_RECALL_{index}\n"),
-            )
-            .unwrap();
-        }
-
-        let seen_system_prompts = Arc::new(std::sync::Mutex::new(Vec::new()));
-        let mut state = create_test_state_with_provider(RecordingToolCallProvider::new(
-            Vec::new(),
-            "ALAN_RECENT_RECALL_4",
-            seen_system_prompts.clone(),
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir);
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text("What did we do yesterday?")]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-
-        let system_prompts = seen_system_prompts.lock().unwrap();
-        let request_prompt = system_prompts
-            .iter()
-            .find(|prompt| prompt.contains("## Runtime Recall Bundle"))
-            .expect("expected runtime recall bundle prompt");
-        assert!(request_prompt.contains("## Runtime Recall Bundle"));
-        assert!(request_prompt.contains("/memory/daily/2026-04-16.md"));
-        assert!(request_prompt.contains("/memory/episodic/2026/04/16/process-4.md"));
-        assert!(request_prompt.contains("ALAN_RECENT_RECALL_4"));
-        assert!(!request_prompt.contains("/memory/topics/recent-match-4.md"));
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_pre_turn_compaction_accounts_for_runtime_recall_budget() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let memory_dir = definition_root.join("memory-store");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        std::fs::write(
-            memory_dir.join("USER.md"),
-            format!(
-                "# User Memory\n- Favorite runtime marker: {}\n",
-                "ALAN_PRETURN_RECALL ".repeat(80)
-            ),
-        )
-        .unwrap();
-
-        let seen_system_prompts = Arc::new(std::sync::Mutex::new(Vec::new()));
-        let mut state = create_test_state_with_provider(RecordingToolCallProvider::new(
-            Vec::new(),
-            "COMPACTED_FOR_RECALL",
-            seen_system_prompts.clone(),
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir.clone());
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-        state.runtime_config.compaction_keep_last = 2;
-        state.runtime_config.compaction_trigger_messages = usize::MAX;
-        state.runtime_config.compaction_soft_trigger_ratio = 1.0;
-        state.runtime_config.compaction_hard_trigger_ratio = 1.0;
-        for idx in 0..3 {
-            state
-                .machine
-                .add_user_message(&format!("Earlier user context {idx} {}", "u".repeat(220)));
-            state.machine.add_assistant_message(
-                &format!("Earlier assistant context {idx} {}", "a".repeat(220)),
-                None,
-            );
-        }
-
-        let user_input = vec![ContentPart::text("What is my favorite runtime marker?")];
-        let turn_recall_bundle = crate::runtime::memory_recall::build_turn_recall_bundle(
-            Some(memory_dir.as_path()),
-            Some(&user_input),
-        );
-        let pending_prompt_tokens =
-            estimate_pending_turn_prompt_tokens(Some(&user_input), turn_recall_bundle.as_deref());
-        assert!(pending_prompt_tokens > 0);
-
-        let base_prompt_tokens = state.machine.tape.estimated_prompt_tokens();
-        state.runtime_config.context_window_tokens =
-            (base_prompt_tokens + pending_prompt_tokens - 1) as u32;
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(user_input),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert_eq!(state.machine.tape.summary(), Some("COMPACTED_FOR_RECALL"));
-
-        let system_prompts = seen_system_prompts.lock().unwrap();
-        assert_eq!(system_prompts.len(), 2);
-        let request_prompt = system_prompts
-            .iter()
-            .find(|prompt| prompt.contains("## Runtime Recall Bundle"))
-            .expect("expected runtime recall bundle prompt");
-        assert!(request_prompt.contains("## Runtime Recall Bundle"));
-        assert!(request_prompt.contains("ALAN_PRETURN_RECALL"));
-        assert_eq!(state.turn_state.drain_deferred_runtime_actions().len(), 1);
-    }
-
-    #[tokio::test]
-    async fn test_run_turn_omits_runtime_recall_bundle_when_memory_disabled() {
-        let temp = tempfile::TempDir::new().unwrap();
-        let definition_root = temp.path().join("repo");
-        let memory_dir = definition_root.join("memory-store");
-        crate::prompts::ensure_memory_store_layout_at(&memory_dir).unwrap();
-        std::fs::write(
-            memory_dir.join("USER.md"),
-            "# User Memory\n- Favorite runtime marker: ALAN_DISABLED_RECALL\n",
-        )
-        .unwrap();
-
-        let seen_system_prompts = Arc::new(std::sync::Mutex::new(Vec::new()));
-        let mut state = create_test_state_with_provider(RecordingToolCallProvider::new(
-            Vec::new(),
-            "ok",
-            seen_system_prompts.clone(),
-        ));
-        state.core_config.memory.store_dir = Some(memory_dir);
-        state.core_config.memory.enabled = false;
-        state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = run_turn_with_cancel(
-            &mut state,
-            TurnRunKind::NewTurn,
-            Some(vec![ContentPart::text(
-                "What is my favorite runtime marker?",
-            )]),
-            &mut emit,
-            &cancel,
-            None,
-        )
-        .await;
-
-        assert!(result.is_ok());
-
-        let system_prompts = seen_system_prompts.lock().unwrap();
-        let request_prompt = system_prompts.last().expect("expected system prompt");
-        assert!(!request_prompt.contains("## Runtime Recall Bundle"));
-        assert!(!request_prompt.contains("ALAN_DISABLED_RECALL"));
-    }
-
-    #[tokio::test]
-    async fn test_maybe_compact_mid_turn_accounts_for_runtime_prompt_overhead() {
-        let mut state = create_test_state_with_provider(ContentMockProvider::new(
-            "MID_TURN_COMPACTION_SUMMARY",
-        ));
-        state.runtime_config.compaction_keep_last = 2;
-        state.runtime_config.compaction_trigger_messages = usize::MAX;
-        state.runtime_config.compaction_soft_trigger_ratio = 1.0;
-        state.runtime_config.compaction_hard_trigger_ratio = 1.0;
-        for idx in 0..3 {
-            state
-                .machine
-                .add_user_message(&format!("Mid-turn user context {idx} {}", "u".repeat(220)));
-            state.machine.add_assistant_message(
-                &format!("Mid-turn assistant context {idx} {}", "a".repeat(220)),
-                None,
-            );
-        }
-
-        let pending_guardrail_instruction = format!(
-            "Retry with a corrected answer and preserve tool intent.\n{}",
-            "guardrail-overhead ".repeat(80)
-        );
-        let additional_prompt_tokens = estimate_request_prompt_overhead_tokens(
-            None,
-            Some(pending_guardrail_instruction.as_str()),
-        );
-        assert!(additional_prompt_tokens > 0);
-
-        let base_prompt_tokens = state.machine.tape.estimated_prompt_tokens();
-        state.runtime_config.context_window_tokens =
-            (base_prompt_tokens + additional_prompt_tokens - 1) as u32;
-
-        let cancel = CancellationToken::new();
-        let mut emit = |_event: Event| async {};
-        let result = maybe_compact_mid_turn_if_needed(
-            &mut state,
-            &mut emit,
-            &cancel,
-            additional_prompt_tokens,
-        )
-        .await;
-
-        assert!(result.is_ok());
-        assert_eq!(
-            state.machine.tape.summary(),
-            Some("MID_TURN_COMPACTION_SUMMARY")
-        );
-        assert_eq!(state.turn_state.compactions_this_turn(), 1);
-    }
+    mod confirmation;
+    mod memory_surfaces;
+    mod recording_provider;
+    mod runtime_recall;
+
+    use recording_provider::RecordingToolCallProvider;
 
     #[tokio::test]
     async fn test_run_turn_resume_turn_preserves_active_skill_context() {
