+++ crates/agent-engine/src/runtime/turn_executor/tests/confirmation.rs
@@ -1,0 +1,85 @@
+use super::*;
+
+#[tokio::test]
+async fn test_run_turn_confirmation_includes_active_skill_permission_hints() {
+    let temp = tempfile::TempDir::new().unwrap();
+    let definition_root = temp.path().join("repo");
+    let skill_dir = definition_root.join("skills/release-check");
+    std::fs::create_dir_all(&skill_dir).unwrap();
+    std::fs::write(
+        skill_dir.join("SKILL.md"),
+        r#"---
+name: Release Check
+description: Review risky release actions
+---
+
+# Instructions
+Use this skill when asked.
+"#,
+    )
+    .unwrap();
+    std::fs::write(
+        skill_dir.join("skill.yaml"),
+        r#"
+runtime:
+  permission_hints:
+    - "May require write approval."
+"#,
+    )
+    .unwrap();
+
+    let mut state = create_test_state_with_provider(ToolCallMockProvider::new(
+        vec![ToolCall {
+            id: Some("call_1".to_string()),
+            name: "request_confirmation".to_string(),
+            arguments: json!({
+                "checkpoint_type": "test",
+                "summary": "Confirm risky action"
+            }),
+        }],
+        "",
+    ));
+    state.prompt_cache = prompt_cache_for_definition_root(&definition_root, Vec::new());
+    let cancel = CancellationToken::new();
+
+    let mut events = vec![];
+    let mut emit = |event: Event| {
+        events.push(event);
+        async {}
+    };
+
+    let result = run_turn_with_cancel(
+        &mut state,
+        TurnRunKind::NewTurn,
+        Some(vec![ContentPart::text(
+            "please use $release-check for this task",
+        )]),
+        &mut emit,
+        &cancel,
+        None,
+    )
+    .await;
+
+    assert!(result.is_ok());
+
+    let confirmation = events.into_iter().find_map(|event| match event {
+        Event::Yield {
+            kind: alan_agent_protocol::YieldKind::Confirmation,
+            payload,
+            ..
+        } => Some(payload),
+        _ => None,
+    });
+    let confirmation = confirmation.expect("expected confirmation yield");
+    let hints = confirmation["details"]["skill_permission_hints"]
+        .as_array()
+        .cloned()
+        .unwrap();
+
+    assert_eq!(hints.len(), 1);
+    assert_eq!(hints[0]["skill_id"], "release-check");
+    assert_eq!(
+        hints[0]["permission_hints"][0],
+        "May require write approval."
+    );
+}
